requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
python-dotenv==1.0.1
Werkzeug==3.0.1
gunicorn==21.2.0
rank-bm25==0.2.2
nltk==3.8.1
scikit-learn==1.4.1
psycopg2-binary==2.9.9 